* nuked DeploySupervisedModel and tests (that were really only testing model fit, nothing deploy specific
