+++ healthcareai/__init__.py
@@ -1,7 +1,5 @@
-from .deploy_supervised_model import DeploySupervisedModel
 from .develop_supervised_model import DevelopSupervisedModel
 
 __all__ = [
-    'DeploySupervisedModel',
     'DevelopSupervisedModel'
 ]
